Time inputs in spins_plot2d adjusted for time limits
Time parameters in spins_plot2d (value in quotes as second argument)
is adjusted to account for the temporal limits of the simulation.
ie. if the simulation only goes until 5s and you as for the output
at t=10s, then this will be adjust to the t=5s output.