Add characterize wave function and new contour solver
A function for characterizing an internal wave has been added.
This function finds the location, amplitude, and the fore and aft
wavelengths of the wave. It uses a specified (or multiple)
isopycnal to calculate these. An associated function for plotting
the results of this script is also added.

Contourcs.m has been replaced with contour_data.m which works with
non-vector grids (ie. mapped coordinates).

rbg.m gives rgb values for names of varies colours. Sounds useful.

A few minor (mostly cosmetic) changes were made to other functions
as well.