Add minor fix for when reader is given incorrect inputs
This fix is largely made so that there can be similar syntax
between 2D and 3D simulations. 3D simulation require 3 arguments
for position, while 2D only needs 2. I still want to be able to
give 3 no matter what type of simluation.

The problem occures if a 2D simulation is given
3 arguemnts, then the second one will very likely be outside
the expected range and will give a 3D variable with an empty
dimension. This is not desired since we want a 2D variable.