Enable plotting of 2D slices from SPINS
SPINS now (at least on a development branch) is able to write
2D slices of a 3D fields at high temporal accuracy. This change
handles the changes associated with those slices which currently
use the same naming convention as full 3D fields (the suffix is
the output number), except that the plane is specified after a dash,
ie. u-xz.5 is the 5th slice of the xz plane of u. Only a single
cross-section per dimension is allowed with this convention until
a new one is defined.