Add filtering option to plot_diagnos
The energy budget is quite chaotic when the resolution is small
relative to the size of the domain. This is most noticeable
when the domain is on a field scale. Smoothing keeps the changes
in the budget, but removes most of the noise. You might need to
play with the smoothing parameters.